+++ bpf/kfunc/samples/cgrp_storage_get.bpf.c
@@ -1,0 +1,9 @@
+/**
+ * void *bpf_cgrp_storage_get(struct bpf_map *map, struct cgroup *cgroup, void *value, u64 flags);
+ */
+#include "vmlinux.h"
+
+void test(void)
+{
+	bpf_cgrp_storage_get(0, 0, 0, 0);
+}